Add the template option. Don't use $PREFIX in the group names.
Added a new required option to createnodes.sh to specify the name of the
KVM template to use during VM cloning.

When saving the hosts to the inventory file in createnodes.sh don't use
the prefix in the name of the groups. Likewise the group names used in
upkub.yml don't include the prefix.

In the ansible playbook the name of the user created in the master is
taken from the play variable "user"