New playbook to join nodes to the kubernetes cluster. Install flannel
A new simple playbook to join nodes to the kubernetes cluster has been
created, it reads the join token from the variable join_token in the
file group_vars/all, you have to add the correct value to this variable.

The playbook key_install now uses a relative path for the key to be
installed.

The playbook upkub.yml uses a relative path for the key to be installed
in the user created in the master node. Also a task to install flannel
has been added